fix(ui): taskContent and select2 working with autoform-select2 4.x
