+++ .meteor/packages
@@ -68,6 +68,7 @@ leaonline:theme
 # FORMS
 #========================================
 aldeed:autoform@8.0.0
+aldeed:autoform-select2@4.0.0-rc.0
 communitypackages:autoform-bootstrap5
 
 # d3k4y:autoform-markdown
